refactor: Merge UserInput and UserOutput class
- Both classes use stdout to print message to user
- Sometimes we need to check about order of stdout
- So, merge two classes into UserIO class
