--- src/main/java/baseball/UserOutput.java
@@ -1,7 +1,0 @@
-package baseball;
-
-public class UserOutput {
-    public void printMessage(String msg) {
-        System.out.print(msg);
-    }
-}
